refactors a little
